% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_zonal_representative.R
\name{assign_zonal_representative}
\alias{assign_zonal_representative}
\title{assign_zonal_representatie}
\usage{
assign_zonal_representative(shapefile, zone_field, parameter, destination)
}
\arguments{
\item{destination}{}
}
\value{

}
\description{
assign_zonal_representatie
}
