% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_zone_map.R
\name{create_zone_map}
\alias{create_zone_map}
\title{create zone map}
\usage{
create_zone_map(path_to_shapefile, zone_field)
}
\arguments{
\item{point_cloud}{}
}
\value{

}
\description{
create zone map
}
