% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_all_timeseries.R
\name{gather_all_timeseries}
\alias{gather_all_timeseries}
\title{gather all timeseries}
\usage{
gather_all_timeseries(destination, parameter)
}
\arguments{
\item{parameter}{}
}
\value{

}
\description{
gather all timeseries
}
