% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_zonal_timeseries.R
\name{identify_zonal_timeseries}
\alias{identify_zonal_timeseries}
\title{indentify_zonal_timeseries}
\usage{
identify_zonal_timeseries(zone_map, destination)
}
\arguments{
\item{destination}{}
}
\value{

}
\description{
indentify_zonal_timeseries
}
