% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_timeseries_df.R
\name{read_all_timeseries_df}
\alias{read_all_timeseries_df}
\title{read all timeseries df}
\usage{
read_all_timeseries_df(destination)
}
\arguments{
\item{destination}{}
}
\value{

}
\description{
read all timeseries df
}
